Refactor tvmaze_api -> api
Refactor tvmaze_exceptions -> exceptions
